Integration and api final details

+++ app/controllers/api/v1/tokens_controller.rb
@@ -1,7 +1,7 @@
 module Api
   module V1
     class TokensController < Api::ApiController
-      before_action :check_token_expiration, only: :create
+      before_action :check_token_expiration, only: :token
 
       # GET api/v1/tokens
       def token
@@ -18,6 +18,8 @@ module Api
       end
 
       def check_token_expiration
+        return generate_token if Token.count == 0
+
         if Token.last.expired?
           Token.last.destroy
